BUG/MINOR: httpclient: consume partly the blocks when necessary
Consume partly the blocks in the httpclient I/O handler when there is
not enough room in the destination buffer for the whole block or when
the block is not contained entirely in the channel's output.

It prevents the I/O handler to be stuck in cases when we need to modify
the buffer with a filter for exemple.

Must be backported in 2.5.

(cherry picked from commit c8f1eb99b4004444c7674d48d967ea3040a62d6e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>