BUG/MINOR: httpclient: process the response when received before the end of the request
A server could reply a response with a shut before the end of the htx
transfer, in this case the httpclient would leave before computing the
received response.

This patch fixes the issue by calling the "process_data" label instead of
the "more" label which don't do the si_shut.

Must be bacported in 2.5.

(cherry picked from commit 1eca8943218eb79bffb3b7dce194b154314795ba)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>