BUG/MEDIUM: mux-h1: only turn CO_FL_ERROR to CS_FL_ERROR with empty ibuf
A connection-level error must not be turned to a stream-level error if there
are still pending data for that stream, otherwise it can cause the truncation
of the last pending data.

This must be backported to affected releases, at least as far as 2.4,
maybe further.

(cherry picked from commit 99bbdbcc21355c8773474274a319086c0d2999d6)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>