BUG/MINOR: httpclient: end callback in applet release
In case an error provokes the release of the applet, we will never call
the end callback of the httpclient.

In the case of a lua script, it would mean that the lua task will never
be waked up after a yield, letting the lua script stuck forever.

Fix the issue by moving the callback from the end of the iohandler to
the applet release function.

Must be backported in 2.5.

(cherry picked from commit eb0d4c40ac0a21944985a2d3ef6cf21cc05b87a8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>