Revert "BUG/MAJOR: mux-pt: Always destroy the backend connection on detach"
This reverts commit 6a8eac0027ac57c9c7a034d97ce0c2afb06d9d60.

On 2.0, the backend connection is always owned by the session. For pure TCP
connection it is not a problem to close it on detach because it will never
be reused. However in legacy HTTP mode, the backend connection is
reusable. So, the connection will still be referenced after a detach,
leading to a crash if it is reused.

Thus the commit is reverted. Note the issue only concernes the legagy HTTP
mode. It is 2.0-specific. There is no upstream commit it.