BUG/MINOR: mux-h2: do not use timeout http-keep-alive on backend side
Commit 15a4733d5 ("BUG/MEDIUM: mux-h2: make use of http-request and
keep-alive timeouts") omitted to check the side of the connection, and
as a side effect, automatically enabled timeouts on idle backend
connections, which is totally contrary to the principle that they
must be autonomous.

This needs to be backported wherever the patch above is backported.

(cherry picked from commit 6ff91e20232821e8b59bb0bbd36ea7e851b42566)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>